rename: 기존 order파일을 payment로 이동 및 변경
